+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Listen for device flow completion without requiring pressing enter by @daimond113
 - Sync scripts repo in background by @daimond113
 - Don't make CAS files read-only on Windows (file removal is disallowed if the file is read-only) by @daimond113 
+- Validate package names are lowercase by @daimond113
 
 ### Performance
 - Clone dependency repos shallowly by @daimond113